change(#109): Change move from Mail to Mails
