+++ tests/Feature/Auth/ForgotPasswordTest.php
@@ -2,7 +2,7 @@
 
 namespace Bambamboole\LaravelCms\Tests\Feature\Auth;
 
-use Bambamboole\LaravelCms\Auth\Mail\ResetPasswordMail;
+use Bambamboole\LaravelCms\Auth\Mails\ResetPasswordMail;
 use Bambamboole\LaravelCms\Auth\Models\User;
 use Bambamboole\LaravelCms\Tests\TestCase;
 use Illuminate\Support\Facades\Mail;
